backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
boto3
numpy
requests
voyageai
pinecone-client
scikit-learn
